+++ SickBeard.py
@@ -503,17 +503,6 @@ class SickGear(object):
 
                 if install_type in ('git', 'source'):
                     popen_list = [sys.executable, sickbeard.MY_FULLNAME]
-                elif install_type == 'win':
-                    if hasattr(sys, 'frozen'):
-                        # c:\dir\to\updater.exe 12345 c:\dir\to\sickbeard.exe
-                        popen_list = [os.path.join(sickbeard.PROG_DIR, 'updater.exe'), str(sickbeard.PID),
-                                      sys.executable]
-                    else:
-                        logger.log(u"Unknown SB launch method, please file a bug report about this", logger.ERROR)
-                        popen_list = [sys.executable, os.path.join(sickbeard.PROG_DIR, 'updater.py'),
-                                      str(sickbeard.PID),
-                                      sys.executable,
-                                      sickbeard.MY_FULLNAME]
 
                 if popen_list:
                     popen_list += sickbeard.MY_ARGS