+++ CHANGES.md
@@ -11,6 +11,7 @@
 * Change Plex notifications to allow authenticated library updates (port from mmccurdy07/Sick-Beard)
 * Change Config/Notifications/Plex logo and description (adapted port from mmccurdy07/Sick-Beard)
 * Add ability for CSS/JS to target a specific page and layout
+* Remove legacy sickbeard updater and build automation code
 
 [develop changelog]